fix solaris 2.8 build, and test12 on solaris 2.9

+++ testsuite/src/mutatee_util_mt.c
@@ -161,7 +161,7 @@ int threads_equal(thread_t a, thread_t b) {
 }
 
 unsigned long thread_int(thread_t a) {
-    return a;
+    return (unsigned long)a;
 }
 
 void schedYield() {
